Scrapper/requirements.txt
# Trendyol Scraping Project Dependencies

# Web Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# HTTP & Async
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
rich>=13.0.0
tenacity>=8.0.0

# Optional: AI Summary (uncomment if needed)
# openai>=1.0.0
# Admin Panel & Visualization
streamlit>=1.30.0
plotly>=5.18.0
psutil>=5.9.0

# Export
openpyxl>=3.1.0
